requirements.txt
flask
web3.py
pyfingerprint
pymongo
pyserial
cryptography